Refactored Secp256k1 class further
