+++ src/did/did-key-resolver.ts
@@ -5,7 +5,7 @@ import { base58btc } from 'multiformats/bases/base58';
 import { Did } from './did.js';
 import { ed25519 } from '../../src/jose/algorithms/signing/ed25519.js';
 import { Encoder } from '../utils/encoder.js';
-import { Secp256k1 } from '../jose/algorithms/signing/secp256k1.js';
+import { Secp256k1 } from '../utils/secp256k1.js';
 import type { KeyMaterial, PublicJwk } from '../jose/types.js';
 
 /**
